[TMP] Bypass -Werror,-Wimplicit-int-float-conversion for now
See https://github.com/KhronosGroup/OpenCL-CTS/issues/855.

+++ test_common/harness/imageHelpers.cpp
@@ -2730,7 +2730,7 @@ void pack_image_pixel(float *srcVector, const cl_image_format *imageFormat,
                 ptr[i] = (int)CONVERT_INT(
                     srcVector[i], MAKE_HEX_FLOAT(-0x1.0p31f, -1, 31),
                     MAKE_HEX_FLOAT(0x1.fffffep30f, 0x1fffffe, 30 - 23),
-                    CL_INT_MAX);
+                    (float)CL_INT_MAX);
             break;
         }
         case CL_UNSIGNED_INT8: {
@@ -2899,7 +2899,7 @@ void pack_image_pixel_error(const float *srcVector,
                     - (cl_long)CONVERT_INT(
                         srcVector[i], MAKE_HEX_FLOAT(-0x1.0p31f, -1, 31),
                         MAKE_HEX_FLOAT(0x1.fffffep30f, 0x1fffffe, 30 - 23),
-                        CL_INT_MAX));
+                        (float)CL_INT_MAX));
             break;
         }
         case CL_UNSIGNED_INT8: {
